SysTick + PendSV conversion complete & successful
